Moved plugin utils back to plugin
